Fix SYN_STREAM of assoc content send order
The SYN_STREAM frames MUST (according to SPDY v3 spec) arrive at the
client before any data of the primary request stream is sent back to the
client to avoid race conditions. This commit achieves this behaviour by
using the same priority for the assoc streams as for the primary stream.
The queue seems to pick SYN_STREAMS prefered of data streams, so this
approach works.