Enabled client-side .map file support
Instead of parsing the html result, which might be error prone and could
result in un-fetched content (thus screwing with measurement when
comparing push/non-push), the client can be passed the -p flag. It then
uses the .map file from the htdocs (the same file used by the server to
determine what files are to be pushed) folder, to determine which files
should be fetched as resources (images, scripts etc.).